@@ -15,7 +15,7 @@ public class Connectivity {
   static final String CONNECTIVITY_WIFI = "wifi";
   static final String CONNECTIVITY_MOBILE = "mobile";
   static final String CONNECTIVITY_ETHERNET = "ethernet";
-  private ConnectivityManager connectivityManager;
+  private final ConnectivityManager connectivityManager;
 
   public Connectivity(ConnectivityManager connectivityManager) {
     this.connectivityManager = connectivityManager;
@@ -44,24 +44,24 @@ public class Connectivity {
 
   @SuppressWarnings("deprecation")
   private String getNetworkTypeLegacy() {
-    // handle type for Android versions less than Android 9
+    // handle type for Android versions less than Android 6
     android.net.NetworkInfo info = connectivityManager.getActiveNetworkInfo();
     if (info == null || !info.isConnected()) {
-      return "none";
+      return CONNECTIVITY_NONE;
     }
     int type = info.getType();
     switch (type) {
       case ConnectivityManager.TYPE_ETHERNET:
-        return "ethernet";
+        return CONNECTIVITY_ETHERNET;
       case ConnectivityManager.TYPE_WIFI:
       case ConnectivityManager.TYPE_WIMAX:
-        return "wifi";
+        return CONNECTIVITY_WIFI;
       case ConnectivityManager.TYPE_MOBILE:
       case ConnectivityManager.TYPE_MOBILE_DUN:
       case ConnectivityManager.TYPE_MOBILE_HIPRI:
-        return "mobile";
+        return CONNECTIVITY_MOBILE;
       default:
-        return "none";
+        return CONNECTIVITY_NONE;
     }
   }
 
